Rename geochem_inverse_optimise to more appropriate name
