+++ synthetic_test.py
@@ -1,22 +1,22 @@
 import matplotlib.pyplot as plt
 import numpy as np
 
-import geochem_inverse_optimize as gio
+import sample_network_unmix as snu
 
 max_val = 1000  # maximum value for synthetic input
 rel_err = 5  # 5% relative error
 
 # Load in Drainage Network
-sample_network, _ = gio.get_sample_graphs(
+sample_network, _ = snu.get_sample_graphs(
     flowdirs_filename="data/synthetic_topo_d8.asc",
     sample_data_filename="data/synthetic_samples.dat",
 )  # Get upstream basins
 
 plt.figure(figsize=(15, 10))  # Visualise network
 plt.title("Sample Network")
-gio.plot_network(sample_network)
+snu.plot_network(sample_network)
 
-areas = gio.get_unique_upstream_areas(sample_network)
+areas = snu.get_unique_upstream_areas(sample_network)
 
 # Pick random values to set each basin
 synth_upst_concs = {sample: max_val * np.random.rand() for sample in areas.keys()}
@@ -25,9 +25,9 @@ synth_upst_concs = {sample: max_val * np.random.rand() for sample in areas.keys(
 input_export_rates = {sample: np.random.rand() for sample in areas.keys()}
 
 # Generate synthetic upstream concentration map
-upst_conc_map = gio.get_upstream_concentration_map(areas, synth_upst_concs)
+upst_conc_map = snu.get_upstream_concentration_map(areas, synth_upst_concs)
 # Predict concentration at downstream observation points
-mixed_synth_down, _ = gio.mix_downstream(
+mixed_synth_down, _ = snu.mix_downstream(
     sample_network, areas, upst_conc_map, export_rates=input_export_rates
 )
 ## Add noise
@@ -36,13 +36,13 @@ mixed_synth_down, _ = gio.mix_downstream(
 #     for sample, value in mixed_synth_down.items()
 # }
 # Set up problem
-problem = gio.SampleNetworkUnmixer(sample_network, use_regularization=False)
+problem = snu.SampleNetworkUnmixer(sample_network, use_regularization=False)
 # Solve problem using synthetic downstream observations as input
 recovered_down, recovered_up = problem.solve(
     mixed_synth_down, solver="ecos", export_rates=input_export_rates
 )
 # Generate recovered upstream concentration map
-recovered_conc_map = gio.get_upstream_concentration_map(areas, recovered_up)
+recovered_conc_map = snu.get_upstream_concentration_map(areas, recovered_up)
 
 # Extract arrays of predicted and `observed' concentrations
 down_preds = [recovered_down[sample] for sample in recovered_down.keys()]
