+++ unmix_mwe.py
@@ -15,14 +15,14 @@ Finally, it visualizes the predicted downstream concentrations and the upstream 
 import matplotlib.pyplot as plt
 import pandas as pd
 
-import geochem_inverse_optimize as gio
+import sample_network_unmix as snu
 
 # Constants
 element = "Mg"  # Set element
 regularizer_strength = 10 ** (-3.0)
 
 # Load sample network
-sample_network, _ = gio.get_sample_graphs(
+sample_network, _ = snu.get_sample_graphs(
     flowdirs_filename="data/d8.asc",
     sample_data_filename="data/sample_data.dat",
 )
@@ -32,30 +32,30 @@ obs_data = pd.read_csv("data/sample_data.dat", delimiter=" ")
 obs_data = obs_data.drop(columns=["Bi", "S"])
 
 plt.figure(figsize=(15, 10))  # Visualise network
-gio.plot_network(sample_network)
+snu.plot_network(sample_network)
 plt.show()
 print("Building problem...")
-problem = gio.SampleNetworkUnmixer(sample_network=sample_network)
+problem = snu.SampleNetworkUnmixer(sample_network=sample_network)
 
 
-element_data = gio.get_element_obs(
+element_data = snu.get_element_obs(
     element, obs_data
 )  # Return dictionary of {sample_name:concentration}
 
-gio.plot_sweep_of_regularizer_strength(problem, element_data, -5, -1, 11)
+snu.plot_sweep_of_regularizer_strength(problem, element_data, -5, -1, 11)
 
 print("Solving problem...")
 element_pred_down, element_pred_upstream = problem.solve(
     element_data, solver="ecos", regularization_strength=10 ** (-3)
 )
 
-area_dict = gio.get_unique_upstream_areas(sample_network)  # Extract areas for each basin
-upstream_map = gio.get_upstream_concentration_map(
+area_dict = snu.get_unique_upstream_areas(sample_network)  # Extract areas for each basin
+upstream_map = snu.get_upstream_concentration_map(
     area_dict, element_pred_upstream
 )  # Assign to upstream preds
 
 # Visualise outputs downstream
-gio.visualise_downstream(pred_dict=element_pred_down, obs_dict=element_data, element=element)
+snu.visualise_downstream(pred_dict=element_pred_down, obs_dict=element_data, element=element)
 plt.show()
 # Visualise outputs upstream
 plt.imshow(upstream_map)
